Remove remnants of paperclip
problem
-------

we've long been on ActiveStorage.  So we don't need to refer to paperclip any more.

solution
-----

remove all references to paperclip
