+++ Gemfile
@@ -31,7 +31,6 @@ gem 'mojo_magick'
 gem 'mysql2'
 gem 'nokogiri'
 gem 'opensearch-ruby'
-gem 'paperclip'
 gem 'postmark'
 gem 'postmark-rails'
 gem 'puma'
